% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eflows.heat.R
\docType{package}
\name{eflows.heat-package}
\alias{eflows.heat}
\alias{eflows.heat-package}
\title{eflows.heat: heat extension to eflows package.}
\description{
eflows.heat: heat extension to eflows package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cvmartin/eflows.heat}
  \item Report bugs at \url{https://github.com/cvmartin/eflows.heat/issues}
}

}
\author{
\strong{Maintainer}: Carlos Varela Martin \email{cvarelamail@gmail.com}

}
