% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{heat_model}
\alias{heat_model}
\title{Generate predictions of inside temperature}
\usage{
heat_model(input_vct, input_ind, hloss, sheat, expand = FALSE)
}
\arguments{
\item{input_vct}{List or DF of vectorized input parameters. `t_out`, `heat_gain`.}

\item{input_ind}{List or DF of parameters with length one `init_t_room`.}

\item{hloss}{heat loss of the building.}

\item{sheat}{specific heat of the building.}

\item{expand}{more detailed resuts.}
}
\description{
Generate predictions of inside temperature
}
